release.py
Add some functionality to determine which AB cluster we are on and set the
base paths to various directories accordingling. The yoctoproject.org cluster
uses /srv/www/vhosts as base. The yocto.io cluster uses /srv/autobuilder. So
now we can guess based on hostname and set this path on the fly.