/release-tools/where_am_i
Use /srv/www/vhosts as the default when we are not on the yocto.io cluster
and on a host that is ambiguous. Odds are that we are on the meta-intel
AB or an older cluster using /srv/www/vhosts instead of /srv/autobuilder.
This facilitates ease of use by meta-intel release script.

Signed-off-by: Tracy Graydon <tracy.graydon@intel.com>