/release-tools/get_cves.py
Takes a revision range and finds the CVEs fixed in between.
Returns the commit has and the commit description from the commit message.
It is not perfect. If there is CVE info in the body of the commit message,
and not in the first line, it will not find it. This is preferable to returning
references to CVE patches that have been dropped, or other extraneous things.

Signed-off-by: Graydon, Tracy <tracy.graydon@intel.com>